muqss: Port "sched: Make nr_iowait_cpu() return 32-bit value"

+++ kernel/sched/MuQSS.c
@@ -3175,7 +3175,7 @@ unsigned long long nr_context_switches(void)
  * runnable.
  */
 
-unsigned long nr_iowait_cpu(int cpu)
+unsigned int nr_iowait_cpu(int cpu)
 {
 	return atomic_read(&cpu_rq(cpu)->nr_iowait);
 }
